Update all utils to be faster

+++ src/main.rs
@@ -7,7 +7,7 @@ mod mangakakalot;
 mod utils;
 
 // imports
-use std::fs;
+use std::fs::{self, DirEntry};
 use tokio;
 use url::Url;
 
@@ -43,16 +43,42 @@ async fn main() -> std::io::Result<()> {
         }
         "compress" => compress().await,
         "clean" => {
+            // check for -y, -F or --force
+            let force = args
+                .iter()
+                .any(|x| x == "-y" || x == "-f" || x == "--force" || x == "-F");
+
+            let exists = fs::metadata("./output").is_ok() || fs::metadata("./zipped").is_ok();
+
+            if !exists {
+                println!("{}{}{}", c.red, "No files to delete", c.end);
+                return Ok(());
+            }
+
+            if !force {
+                println!(
+                    "{}Are you sure you want to delete all files? (y/n){}",
+                    c.red, c.end
+                );
+                let mut input = String::new();
+
+                std::io::stdin().read_line(&mut input).unwrap();
+                if input.trim() != "y" {
+                    println!("{}{}{}", c.red, "Aborting", c.end);
+                    return Ok(());
+                }
+            }
+
             fs::remove_dir_all("./output")?;
             fs::remove_dir_all("./zipped")?;
         }
         _ => {
             download().await.unwrap();
 
-            if args.contains(&"--autocompress".to_string())
-                || globals::Globals::new().auto_compress
-                || args.contains(&"-a".to_string())
-            {
+            let autocompress = args.iter().any(|x| x == "--autocompress" || x == "-a")
+                || Globals::new().auto_compress;
+
+            if autocompress {
                 compress().await;
             }
         }
@@ -73,10 +99,8 @@ async fn download() -> std::io::Result<()> {
         println!("{}{}{}", c.red, "No url given", c.end);
         return Ok(());
     }
-    let url = &args[1];
-    let parsable = Url::parse(url).unwrap();
+    let url = Url::parse(&args[1]).unwrap();
 
-    // search args for --skip [number]
     let mut iter = args.iter();
 
     let mut skip = 0;
@@ -90,22 +114,20 @@ async fn download() -> std::io::Result<()> {
     }
 
     // split url into parts
-    let site_name = parsable.host_str().unwrap();
+    let site_name = url.host_str().unwrap();
 
     fs::create_dir_all(&g.output_dir)?;
 
     match site_name {
-        "mangakakalot.com" => mangakakalot::downloader(parsable, skip).await.unwrap(),
-        "chapmanganato.com" => chapmanganato::downloader(parsable, skip).await.unwrap(),
+        "mangakakalot.com" => mangakakalot::downloader(url, skip).await.unwrap(),
+        "chapmanganato.com" => chapmanganato::downloader(url, skip).await.unwrap(),
         _ => {
             println!("{}{}{}{}", c.red, site_name, " is not supported", c.end);
             return Ok(());
         }
     }
 
-    // println!("{}{}{}{}", "Site name: ", c.green, site_name, c.end);
     let duration = start.elapsed();
-    // println!("{}", duration.as_secs());
 
     println!(
         "{}{} ({}{} seconds) {}",
@@ -119,52 +141,23 @@ async fn download() -> std::io::Result<()> {
 }
 
 fn help() {
-    let help_msg = r#"Usage: ./rust-mangakakalot [command] [url] [options]
-
-Commands:
-    download    Download manga from url
-    compress    Compress downloaded manga into zip files
-    help        Show this message
-
-Options:
-    -l, --list                      List chapters
-    -f, --format                    Set the format of the zip file (default: .cbz)
-    -h, --help                      Show this message
-    -a, --autocompress              Automatically compress downloaded manga into zip files
-    -s [number], --skip [number]    Start downloading from chapter [number]
-    -c [n] or --chapter [n]         Download chapter by index (see --list)
-    -n [n] or --name [n]            Download chapter by name in url (see --list)
-    -r [n] [n], --range [n] [n]     Download chapters from [n] to [n]
-
-    -v, --verbose                   Show detailed information about the download process
-"#;
+    let c = Color::new();
+    let help_msg = format!(include_str!("./help.txt"), c.red, c.end);
 
     println!("{}", help_msg);
 }
 
 async fn compress() {
+    let force = std::env::args().any(|x| x == "--force" || x == "-F");
+
     fs::create_dir_all("./zipped").expect("Failed to create \"zipped\" directory");
+
     let books = fs::read_dir("./output").unwrap();
     for book in books {
-        println!(
-            "Compressing {}",
-            book.as_ref().unwrap().path().to_str().unwrap()
-        );
+        println!("\nCompressing {}", path_replacer(&book).unwrap());
 
-        fs::create_dir_all(
-            "./zipped/".to_string()
-                + &book
-                    .as_ref()
-                    .unwrap()
-                    .path()
-                    .to_str()
-                    .unwrap()
-                    .replace("\\", "/")
-                    .split("/")
-                    .last()
-                    .unwrap(),
-        )
-        .expect("Failed to create \"zipped\" directory");
+        fs::create_dir_all(format!("./zipped/{}", get_last(&book).unwrap()))
+            .expect("Failed to create \"zipped\" directory");
 
         let chapters = fs::read_dir(book.as_ref().unwrap().path()).unwrap();
         for chapter in chapters {
@@ -174,34 +167,58 @@ async fn compress() {
             if re.is_match(&chapter.as_ref().unwrap().path().to_str().unwrap()) {
                 continue;
             }
-            // split / get last
-            compress::compress(
-                &chapter.as_ref().unwrap().path().to_str().unwrap(),
-                &format!(
-                    "./zipped/{}/{}{}",
-                    book.as_ref()
-                        .unwrap()
-                        .path()
-                        .to_str()
-                        .unwrap()
-                        .replace("\\", "/")
-                        .split("/")
-                        .last()
-                        .unwrap(),
-                    &chapter
-                        .as_ref()
-                        .unwrap()
-                        .path()
-                        .to_str()
-                        .unwrap()
-                        .replace("\\", "/")
-                        .split("/")
-                        .last()
-                        .unwrap(),
-                    globals::Globals::new().zip_format
-                ),
-            )
-            .await;
+
+            let path = format!(
+                "{}{}",
+                path_replacer(&chapter).unwrap().replace("output", "zipped"),
+                globals::Globals::new().zip_format
+            );
+
+            println!("\nCompressing {}", path_replacer(&chapter).unwrap());
+
+            if !force && std::path::Path::new(&path).exists() {
+                // get the size of both the zip and the input folder
+                let zip_size = std::fs::metadata(&path).unwrap().len();
+                let mut input_size: u64 = 0;
+
+                for file in fs::read_dir(chapter.as_ref().unwrap().path()).unwrap() {
+                    input_size += std::fs::metadata(file.unwrap().path()).unwrap().len();
+                }
+
+                if zip_size > input_size {
+                    println!("{} already exists, skipping", path);
+                    continue;
+                }
+            }
+
+            compress::compress(&chapter.as_ref().unwrap().path().to_str().unwrap(), &path).await;
         }
+    }
+}
+
+fn get_last(path: &Result<DirEntry, std::io::Error>) -> Result<String, &'static str> {
+    match path {
+        Ok(entry) => {
+            let path_ = entry.path();
+            let path_str = path_.to_str().unwrap();
+            Ok(path_str
+                .replace("\\", "/")
+                .split("/")
+                .last()
+                .unwrap()
+                .to_string())
+        }
+        Err(_) => Err("Error reading path"),
+    }
+}
+
+fn path_replacer(path: &Result<DirEntry, std::io::Error>) -> Result<String, &'static str> {
+    match path {
+        Ok(entry) => {
+            let path_ = entry.path();
+            let path_str = path_.to_str().unwrap();
+            Ok(path_str.replace("\\", "/"))
+        }
+        Err(_) => Err("Error reading path"),
     }
 }
